Tsvetislav/responsefeature

See merge request I476232/s3-chatbot!2

+++ src/main/java/sem3chatbot/backend/business/InputProcessingService.java
@@ -6,5 +6,5 @@ import sem3chatbot.backend.domain.UserInput;
 import java.util.HashMap;
 
 public interface InputProcessingService {
-    HashMap<Integer, String> findMatches(UserInput request);
+    HashMap<Long, String> findMatches(UserInput request);
 }
